+++ packages/app-page-builder/src/admin/components/Table/Table/Actions/EditPage.tsx
@@ -1,31 +1,18 @@
-import React, { useMemo } from "react";
+import React from "react";
 import { ReactComponent as Edit } from "@material-design-icons/svg/outlined/edit.svg";
-import { useFolders } from "@webiny/app-aco";
 import { PageListConfig } from "~/admin/config/pages";
 import { usePage } from "~/admin/views/Pages/hooks/usePage";
 import { useCreatePageFrom } from "~/admin/views/Pages/hooks/useCreatePageFrom";
 import { useNavigatePage } from "~/admin/hooks/useNavigatePage";
-import { usePagesPermissions } from "~/hooks/permissions";
 
 export const EditPage = () => {
     const { page } = usePage();
-    const { folderLevelPermissions: flp } = useFolders();
-    const { canUpdate: pagesCanUpdate } = usePagesPermissions();
     const { OptionsMenuItem, OptionsMenuLink } = PageListConfig.Browser.PageAction;
     const { getPageEditorUrl, navigateToPageEditor } = useNavigatePage();
     const { createPageForm, loading } = useCreatePageFrom({
         page,
         onSuccess: () => navigateToPageEditor(page.data.pid)
     });
-
-    const { folderId } = page.location;
-    const canEdit = useMemo(() => {
-        return pagesCanUpdate(page.data.createdBy.id) && flp.canManageContent(folderId);
-    }, [flp, folderId]);
-
-    if (!canEdit) {
-        return null;
-    }
 
     if (page.data.locked) {
         return (
