+++ packages/app-page-builder/src/admin/components/Table/Table/Actions/MovePage.tsx
@@ -1,6 +1,5 @@
-import React, { useMemo } from "react";
+import React from "react";
 import { ReactComponent as Move } from "@material-design-icons/svg/outlined/drive_file_move.svg";
-import { useFolders } from "@webiny/app-aco";
 import { PageListConfig } from "~/admin/config/pages";
 import { usePage } from "~/admin/views/Pages/hooks/usePage";
 import { useMovePageToFolder } from "~/admin/views/Pages/hooks/useMovePageToFolder";
@@ -9,17 +8,6 @@ export const MovePage = () => {
     const { page } = usePage();
     const movePageToFolder = useMovePageToFolder({ record: page });
     const { OptionsMenuItem } = PageListConfig.Browser.PageAction;
-
-    const { folderLevelPermissions: flp } = useFolders();
-
-    const { folderId } = page.location;
-    const canMove = useMemo(() => {
-        return flp.canManageContent(folderId);
-    }, [flp, folderId]);
-
-    if (!canMove) {
-        return null;
-    }
 
     return (
         <OptionsMenuItem
