+++ packages/app-page-builder/src/admin/components/Table/Table/Actions/index.ts
@@ -1,5 +1,9 @@
 export * from "./ChangePageStatus";
+export * from "./SecureChangePageStatus";
 export * from "./DeletePage";
+export * from "./SecureDeletePage";
 export * from "./EditPage";
+export * from "./SecureEditPage";
 export * from "./MovePage";
+export * from "./SecureMovePage";
 export * from "./PreviewPage";
